packages/telegram/crawl/requirements.txt
python-dotenv==0.21.0
Telethon==1.26.0
asyncio==3.4.3